# thermalforge-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThermalForge API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tasks/{task_id}/thermal-analysis": {
      "get": {
        "tags": [
          "thermal-design"
        ],
        "summary": "Get Thermal Analysis",
        "operationId": "get_thermal_analysis_v1_tasks__task_id__thermal_analysis_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThermalAnalysisResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/thermal-design": {
      "get": {
        "tags": [
          "thermal-design"
        ],
        "summary": "Get Thermal Design",
        "operationId": "get_thermal_design_v1_tasks__task_id__thermal_design_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThermalDesignSpec"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tasks/{task_id}/viewer-manifest": {
      "get": {
        "tags": [
          "viewer"
        ],
        "summary": "Get Viewer Manifest",
        "operationId": "get_viewer_manifest_v1_tasks__task_id__viewer_manifest_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewerManifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/models/{artifact_id}/content": {
      "get": {
        "tags": [
          "viewer"
        ],
        "summary": "Download Viewer Model",
        "operationId": "download_viewer_model_v1_tasks__task_id__models__artifact_id__content_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable model artifact bytes.",
            "content": {
              "application/octet-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/viewer-library/{model_id}/content": {
      "get": {
        "tags": [
          "viewer"
        ],
        "summary": "Download Viewer Library Model",
        "operationId": "download_viewer_library_model_v1_viewer_library__model_id__content_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable curated viewer model bytes.",
            "content": {
              "model/gltf-binary": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisAssumption": {
        "properties": {
          "key": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Key"
          },
          "value": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Value"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason"
          },
          "impact": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Impact"
          },
          "requires_confirmation": {
            "type": "boolean",
            "title": "Requires Confirmation",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "value",
          "reason",
          "impact"
        ],
        "title": "AnalysisAssumption"
      },
...
      "CandidateEvaluation": {
        "properties": {
          "solution_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Solution Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "eligible": {
            "type": "boolean",
            "title": "Eligible"
          },
          "rejection_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rejection Codes"
          },
          "thermal_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Thermal Score"
          },
          "cost_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Cost Score"
          },
          "risk_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score"
          },
          "added_mass_g": {
            "type": "number",
            "minimum": 0.0,
            "title": "Added Mass G"
          },
          "added_mass_percent": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Added Mass Percent"
          },
          "max_temperature_c": {
            "type": "number",
            "title": "Max Temperature C"
          },
          "hotspot_reduction_c": {
            "type": "number",
            "title": "Hotspot Reduction C"
          },
          "interference_risk": {
            "type": "string",
            "enum": [
              "低",
              "中",
              "高"
            ],
            "title": "Interference Risk"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "solution_id",
          "title",
          "eligible",
          "rejection_codes",
          "thermal_score",
          "cost_score",
          "risk_score",
          "added_mass_g",
          "added_mass_percent",
          "max_temperature_c",
          "hotspot_reduction_c",
          "interference_risk"
        ],
        "title": "CandidateEvaluation"
      },
      "CandidateResult": {
        "properties": {
          "curve": {
            "items": {
              "$ref": "#/components/schemas/ThermalCurvePoint"
            },
            "type": "array",
            "maxItems": 5001,
            "minItems": 1,
            "title": "Curve"
          },
          "maxTemperatureC": {
            "type": "number",
            "title": "Maxtemperaturec"
          },
          "timeToLimitMinutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timetolimitminutes"
          },
          "thermalResistanceKPerW": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Thermalresistancekperw"
          },
          "effectiveCapacityJPerK": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Effectivecapacityjperk"
          },
          "solutionId": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Solutionid"
          },
          "score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "grade": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C",
              "D"
            ],
            "title": "Grade"
          },
          "addedMassPercent": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Addedmasspercent"
          },
          "interferenceRisk": {
            "type": "string",
            "enum": [
              "低",
              "中",
              "高"
            ],
            "title": "Interferencerisk"
          },
          "hotspotReductionC": {
            "type": "number",
            "title": "Hotspotreductionc"
          },
          "timeToLimitImprovementPercent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timetolimitimprovementpercent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "curve",
          "maxTemperatureC",
          "timeToLimitMinutes",
          "thermalResistanceKPerW",
          "effectiveCapacityJPerK",
          "solutionId",
          "score",
          "grade",
          "addedMassPercent",
          "interferenceRisk",
          "hotspotReductionC",
          "timeToLimitImprovementPercent"
        ],
        "title": "CandidateResult"
      },
...
      "ComponentExplanation": {
        "properties": {
          "component_id": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]{0,79}$",
            "title": "Component Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "explanation": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Explanation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "component_id",
          "name",
          "explanation"
        ],
        "title": "ComponentExplanation"
      },
      "DesignRisk": {
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "engineering_brief",
              "thermal_analysis",
              "solution_catalog",
              "llm_assessment"
            ],
            "title": "Source"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Description"
          },
          "impact": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Impact"
          },
          "recommended_action": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Recommended Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "description",
          "impact",
          "recommended_action"
        ],
        "title": "DesignRisk"
      },
...
      "GenerationBrief": {
        "properties": {
          "design_intent": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Design Intent"
          },
          "positive_constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Positive Constraints"
          },
          "negative_constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Negative Constraints"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "design_intent",
          "positive_constraints"
        ],
        "title": "GenerationBrief"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SelectedThermalSolution": {
        "properties": {
          "solution_id": {
            "type": "string",
            "title": "Solution Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          },
          "score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "grade": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C",
              "D"
            ],
            "title": "Grade"
          },
          "max_temperature_c": {
            "type": "number",
            "title": "Max Temperature C"
          },
          "time_to_limit_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time To Limit Minutes"
          },
          "thermal_resistance_k_per_w": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Thermal Resistance K Per W"
          },
          "effective_capacity_j_per_k": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Effective Capacity J Per K"
          },
          "added_mass_g": {
            "type": "number",
            "minimum": 0.0,
            "title": "Added Mass G"
          },
          "added_mass_percent": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Added Mass Percent"
          },
          "interference_risk": {
            "type": "string",
            "enum": [
              "低",
              "中",
              "高"
            ],
            "title": "Interference Risk"
          },
          "hotspot_reduction_c": {
            "type": "number",
            "title": "Hotspot Reduction C"
          },
          "time_to_limit_improvement_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time To Limit Improvement Percent"
          },
          "cost_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Cost Score"
          },
          "risk_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "solution_id",
          "title",
          "tag",
          "features",
          "score",
          "grade",
          "max_temperature_c",
          "time_to_limit_minutes",
          "thermal_resistance_k_per_w",
          "effective_capacity_j_per_k",
          "added_mass_g",
          "added_mass_percent",
          "interference_risk",
          "hotspot_reduction_c",
          "time_to_limit_improvement_percent",
          "cost_score",
          "risk_score"
        ],
        "title": "SelectedThermalSolution"
      },
...
      "TaskRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "stage": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "status",
          "stage",
          "prompt",
          "idempotency_key",
          "created_at",
          "updated_at"
        ],
        "title": "TaskRead"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "created",
          "uploaded",
          "parsing",
          "awaiting_input",
          "briefing",
          "thermal_analysis",
          "concept_imaging",
          "multiview_imaging",
          "multiview_review",
          "modeling",
          "model_review",
          "ready",
          "failed",
          "cancelled"
        ],
        "title": "TaskStatus"
      },
      "ThermalAnalysisResult": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Id"
          },
          "generatedAt": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Generatedat"
          },
          "source": {
            "type": "string",
            "enum": [
              "engineering-estimate",
              "measured-calibrated"
            ],
            "title": "Source"
          },
          "methodLabel": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Methodlabel"
          },
          "totalPowerW": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Totalpowerw"
          },
          "baseline": {
            "$ref": "#/components/schemas/ThermalCaseResult"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateResult"
            },
            "type": "array",
            "minItems": 1,
            "title": "Candidates"
          },
          "recommendedSolutionId": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Recommendedsolutionid"
          },
          "riskLevel": {
            "type": "string",
            "enum": [
              "Low",
              "Medium",
              "High"
            ],
            "title": "Risklevel"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "generatedAt",
          "source",
          "methodLabel",
          "totalPowerW",
          "baseline",
          "candidates",
          "recommendedSolutionId",
          "riskLevel",
          "warnings"
        ],
        "title": "ThermalAnalysisResult"
      },
      "ThermalCaseResult": {
        "properties": {
          "curve": {
            "items": {
              "$ref": "#/components/schemas/ThermalCurvePoint"
            },
            "type": "array",
            "maxItems": 5001,
            "minItems": 1,
            "title": "Curve"
          },
          "maxTemperatureC": {
            "type": "number",
            "title": "Maxtemperaturec"
          },
          "timeToLimitMinutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timetolimitminutes"
          },
          "thermalResistanceKPerW": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Thermalresistancekperw"
          },
          "effectiveCapacityJPerK": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Effectivecapacityjperk"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "curve",
          "maxTemperatureC",
          "timeToLimitMinutes",
          "thermalResistanceKPerW",
          "effectiveCapacityJPerK"
        ],
        "title": "ThermalCaseResult"
      },
...
      "ThermalDesignSpec": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "engineering_brief_artifact_id": {
            "type": "string",
            "title": "Engineering Brief Artifact Id"
          },
          "thermal_analysis_artifact_id": {
            "type": "string",
            "title": "Thermal Analysis Artifact Id"
          },
          "analysis_id": {
            "type": "string",
            "title": "Analysis Id"
          },
          "baseline_max_temperature_c": {
            "type": "number",
            "title": "Baseline Max Temperature C"
          },
          "baseline_time_to_limit_minutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Time To Limit Minutes"
          },
          "selected_solution": {
            "$ref": "#/components/schemas/SelectedThermalSolution"
          },
          "candidate_evaluations": {
            "items": {
              "$ref": "#/components/schemas/CandidateEvaluation"
            },
            "type": "array",
            "title": "Candidate Evaluations"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "heat_transfer_path": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Heat Transfer Path"
          },
          "material_recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Material Recommendations"
          },
          "geometry_anchors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Geometry Anchors"
          },
          "manufacturing_recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Manufacturing Recommendations"
          },
          "component_explanations": {
            "items": {
              "$ref": "#/components/schemas/ComponentExplanation"
            },
            "type": "array",
            "title": "Component Explanations"
          },
          "generation_brief": {
            "$ref": "#/components/schemas/GenerationBrief"
          },
          "assumptions": {
            "items": {
              "$ref": "#/components/schemas/AnalysisAssumption"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/DesignRisk"
            },
            "type": "array",
            "title": "Risks"
          },
          "unverified_items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unverified Items"
          },
          "requires_human_confirmation": {
            "type": "boolean",
            "title": "Requires Human Confirmation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "engineering_brief_artifact_id",
          "thermal_analysis_artifact_id",
          "analysis_id",
          "baseline_max_temperature_c",
          "baseline_time_to_limit_minutes",
          "selected_solution",
          "candidate_evaluations",
          "rationale",
          "heat_transfer_path",
          "material_recommendations",
          "geometry_anchors",
          "manufacturing_recommendations",
          "component_explanations",
          "generation_brief",
          "assumptions",
          "risks",
          "unverified_items",
          "requires_human_confirmation"
        ],
        "title": "ThermalDesignSpec"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ViewerManifest": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "asset": {
            "$ref": "#/components/schemas/ViewerAsset"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/ViewerVariant"
            },
            "type": "array",
            "title": "Variants"
          },
          "notices": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notices"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "asset"
        ],
        "title": "ViewerManifest"
      },
...
    }
  }
}